chore: move rebuild to core-utils
