+++ packages/api/core/src/api/start.ts
@@ -1,6 +1,6 @@
 import { spawn, SpawnOptions } from 'child_process';
 
-import { getElectronVersion } from '@electron-forge/core-utils';
+import { getElectronVersion, listrCompatibleRebuildHook } from '@electron-forge/core-utils';
 import { ElectronProcess, ForgeArch, ForgePlatform, ResolvedForgeConfig, StartOptions } from '@electron-forge/shared-types';
 import chalk from 'chalk';
 import debug from 'debug';
@@ -10,7 +10,6 @@ import locateElectronExecutable from '../util/electron-executable';
 import getForgeConfig from '../util/forge-config';
 import { runHook } from '../util/hook';
 import { readMutatedPackageJson } from '../util/read-package-json';
-import { listrCompatibleRebuildHook } from '../util/rebuild';
 import resolveDir from '../util/resolve-dir';
 
 const d = debug('electron-forge:start');
