+++ packages/api/core/test/fast/start_spec.ts
@@ -23,6 +23,7 @@ describe('start', () => {
       '../util/electron-executable': () => Promise.resolve('fake_electron_path'),
       '@electron-forge/core-utils': {
         getElectronVersion: () => Promise.resolve('1.0.0'),
+        listrCompatibleRebuildHook: () => Promise.resolve(),
       },
       '../util/forge-config': async () => ({
         pluginInterface: {
@@ -33,9 +34,6 @@ describe('start', () => {
       '../util/resolve-dir': async (dir: string) => resolveStub(dir),
       '../util/read-package-json': {
         readMutatedPackageJson: () => Promise.resolve(packageJSON),
-      },
-      '../util/rebuild': {
-        listrCompatibleRebuildHook: () => Promise.resolve(),
       },
       child_process: {
         spawn: spawnStub,
